FISH-468 Use ConfigExtension and change package name
