+++ enterprise-backwards-compatibility/backwards-compatibility-core/src/main/java/fish/payara/extensions/notifiers/compat/UpgradeService.java
@@ -37,7 +37,7 @@
  * only if the new code is made subject to such option by the copyright
  * holder.
  */
-package fish.payara.extensions.notifiers;
+package fish.payara.extensions.notifiers.compat;
 
 import com.sun.enterprise.config.serverbeans.Config;
 import com.sun.enterprise.config.serverbeans.Configs;
