+++ enterprise-backwards-compatibility/xmpp-notifier-backwards-compatibility/src/main/java/fish/payara/extensions/notifiers/xmpp/compat/XmppHealthCheckNotifierConfigurer.java
@@ -37,11 +37,11 @@
  * only if the new code is made subject to such option by the copyright
  * holder.
  */
-package fish.payara.extensions.notifiers.snmp;
+package fish.payara.extensions.notifiers.xmpp.compat;
 
 
 import fish.payara.extensions.notifiers.BaseSetNotifierConfigurationCommand;
-import fish.payara.jmx.monitoring.configuration.MonitoringServiceConfiguration;
+import fish.payara.nucleus.healthcheck.configuration.HealthCheckServiceConfiguration;
 import org.glassfish.api.admin.AdminCommandContext;
 import org.glassfish.api.admin.ExecuteOn;
 import org.glassfish.api.admin.RestEndpoint;
@@ -53,28 +53,25 @@ import org.glassfish.hk2.api.PerLookup;
 import org.jvnet.hk2.annotations.Service;
 
 /**
- * Asadmin command to configure the SNMP notifier with the monitoring service.
- * Deprecated, folded into {@link fish.payara.jmx.monitoring.admin.SetJMXMonitoringConfiguration}
- * @since 4.1.2.174
- * @author jonathan coustick
+ * @author mertcaliskan
+ * @deprecated folded into {@link fish.payara.nucleus.healthcheck.admin.SetHealthCheckConfiguration}
  */
 @Deprecated
-@Service(name = "monitoring-snmp-notifier-configure")
+@Service(name = "healthcheck-xmpp-notifier-configure")
 @PerLookup
 @ExecuteOn({RuntimeType.DAS, RuntimeType.INSTANCE})
 @TargetType(value = {CommandTarget.DAS, CommandTarget.STANDALONE_INSTANCE, CommandTarget.CLUSTER, CommandTarget.CLUSTERED_INSTANCE, CommandTarget.CONFIG})
 @RestEndpoints({
-        @RestEndpoint(configBean = MonitoringServiceConfiguration.class,
+        @RestEndpoint(configBean = HealthCheckServiceConfiguration.class,
                 opType = RestEndpoint.OpType.POST,
-                path = "monitoring-snmp-notifier-configure",
-                description = "Configures SNMP Notifier for Monitoring Service")
+                path = "healthcheck-xmpp-notifier-configure",
+                description = "Configures XMPP Notifier for HealthCheck Service")
 })
-public class SnmpMonitoringNotifierConfigurer extends BaseSetNotifierConfigurationCommand {
+public class XmppHealthCheckNotifierConfigurer extends BaseSetNotifierConfigurationCommand {
 
     @Override
-    public void execute(final AdminCommandContext context) {
-        configureNotifier(context, "set-snmp-notifier-configuration");
-        configureService(context, "set-jmx-monitoring-configuration", "snmp-notifier");
+    public void execute(AdminCommandContext context) {
+        configureNotifier(context, "set-xmpp-notifier-configuration");
+        configureService(context, "set-healthcheck-configuration", "xmpp-notifier");
     }
-
 }
